+++ integration_tests/suite/helpers/base.py
@@ -146,10 +146,12 @@ class IntegrationTest(AssetLaunchingTestCase):
     @contextmanager
     def confd_stopped(cls):
         cls.stop_service('confd')
-        yield
-        cls.start_service('confd')
-        cls.reset_clients()
-        until.true(cls.confd.is_up, tries=5)
+        try:
+            yield
+        finally:
+            cls.start_service('confd')
+            cls.reset_clients()
+            until.true(cls.confd.is_up, tries=5)
 
 
 class RealAsteriskIntegrationTest(IntegrationTest):