+++ xivo_ctid_ng/plugins/conferences/exceptions.py
@@ -18,16 +18,32 @@ class NoSuchConference(APIException):
         )
 
 
-class ParticipantListError(APIException):
-    def __init__(self, conference_id, tenant_uuid, message):
+class NoSuchParticipant(APIException):
+    def __init__(self, tenant_uuid, conference_id, participant_id):
+        super().__init__(
+            status_code=404,
+            message='Conference id "{}" has no such participant: id "{}"'.format(conference_id, participant_id),
+            error_id='no-such-participant',
+            resource='conference-participant',
+            details={
+                'tenant_uuid': tenant_uuid,
+                'conference_id': conference_id,
+                'participant_id': participant_id,
+            }
+        )
+
+
+class ConferenceParticipantError(APIException):
+    def __init__(self, conference_id, tenant_uuid, message, participant_id=None):
         super().__init__(
             status_code=500,
-            message='Error while listing participants of conference "{}": "{}"'.format(conference_id, message),
-            error_id='participant-list-error',
-            resource='conference',
+            message='Error while operating on participants of conference "{}": "{}"'.format(conference_id, message),
+            error_id='conference-participant-error',
+            resource='conference-participant',
             details={
                 'conference_id': conference_id,
                 'tenant_uuid': tenant_uuid,
                 'original_message': message,
+                'participant_id': participant_id,
             }
         )