+++ xivo_ctid_ng/plugins/conferences/plugin.py
@@ -4,7 +4,10 @@
 from xivo_amid_client import Client as AmidClient
 from xivo_confd_client import Client as ConfdClient
 
-from .resources import ParticipantsResource
+from .resources import (
+    ParticipantsResource,
+    ParticipantResource,
+)
 from .services import ConferencesService
 
 
@@ -12,6 +15,7 @@ class Plugin:
 
     def load(self, dependencies):
         api = dependencies['api']
+        ari = dependencies['ari']
         config = dependencies['config']
         token_changed_subscribe = dependencies['token_changed_subscribe']
 
@@ -21,6 +25,7 @@ class Plugin:
         token_changed_subscribe(amid_client.set_token)
         token_changed_subscribe(confd_client.set_token)
 
-        conferences_service = ConferencesService(amid_client, confd_client)
+        conferences_service = ConferencesService(amid_client, ari.client, confd_client)
 
         api.add_resource(ParticipantsResource, '/conferences/<int:conference_id>/participants', resource_class_args=[conferences_service])
+        api.add_resource(ParticipantResource, '/conferences/<int:conference_id>/participants/<participant_id>', resource_class_args=[conferences_service])