+++ xivo_ctid_ng/plugins/conferences/resources.py
@@ -23,3 +23,15 @@ class ParticipantsResource(AuthResource):
             'total': len(participants),
         }
         return items, 200
+
+
+class ParticipantResource(AuthResource):
+
+    def __init__(self, conferences_service):
+        self._service = conferences_service
+
+    @required_acl('ctid-ng.conferences.{conference_id}.participants.{participant_id}.delete')
+    def delete(self, conference_id, participant_id):
+        tenant = Tenant.autodetect()
+        self._service.kick_participant(tenant.uuid, conference_id, participant_id)
+        return '', 204