+++ xivo_ctid_ng/plugins/conferences/services.py
@@ -3,6 +3,7 @@
 
 import logging
 
+from ari.exceptions import ARINotFound
 from marshmallow import ValidationError
 from requests import RequestException
 from xivo_ctid_ng.exceptions import (
@@ -12,7 +13,8 @@ from xivo_ctid_ng.exceptions import (
 
 from .exceptions import (
     NoSuchConference,
-    ParticipantListError,
+    NoSuchParticipant,
+    ConferenceParticipantError,
 )
 from .schemas import participant_schema
 
@@ -21,8 +23,9 @@ logger = logging.getLogger(__name__)
 
 class ConferencesService:
 
-    def __init__(self, amid, confd):
+    def __init__(self, amid, ari, confd):
         self._amid = amid
+        self._ari = ari
         self._confd = confd
 
     def list_participants(self, conference_id, tenant_uuid):
@@ -45,7 +48,7 @@ class ConferencesService:
 
         if participant_list_result['Response'] != 'Success':
             message = participant_list_result['Message']
-            raise ParticipantListError(conference_id, tenant_uuid, message)
+            raise ConferenceParticipantError(conference_id, tenant_uuid, message)
 
         result = []
         for participant_list_item in participant_list:
@@ -65,7 +68,36 @@ class ConferencesService:
             try:
                 participant = participant_schema.load(raw_participant).data
             except ValidationError as e:
-                raise ParticipantListError(conference_id, tenant_uuid, str(e))
+                raise ConferenceParticipantError(conference_id, tenant_uuid, str(e))
             result.append(participant)
 
         return result
+
+    def kick_participant(self, tenant_uuid, conference_id, participant_id):
+        try:
+            conferences = self._confd.conferences.list(tenant_uuid=tenant_uuid, recurse=True)['items']
+        except RequestException as e:
+            raise XiVOConfdUnreachable(self._confd, e)
+
+        if conference_id not in (conference['id'] for conference in conferences):
+            raise NoSuchConference(conference_id, tenant_uuid)
+
+        participants = self.list_participants(conference_id, tenant_uuid)
+        if participant_id not in [participant['id'] for participant in participants]:
+            raise NoSuchParticipant(tenant_uuid, conference_id, participant_id)
+
+        try:
+            channel = self._ari.channels.get(channelId=participant_id)
+        except ARINotFound:
+            raise NoSuchParticipant(tenant_uuid, conference_id, participant_id)
+
+        try:
+            response_items = self._amid.action('ConfbridgeKick', {'conference': conference_id,
+                                                                  'channel': channel.json['name']})
+        except RequestException as e:
+            raise XiVOAmidError(self._amid, e)
+
+        response = response_items[0]
+        if response['Response'] != 'Success':
+            message = response['Message']
+            raise ConferenceParticipantError(conference_id, tenant_uuid, message, participant_id)